Refactor Manage Bands screens
